\name{ydays}
\alias{ydays}
\alias{years}
\title{Extractor functions for POSIXct objects}
\description{Like \code{months} but returning years and days-of-year (starting at 1).
}
\usage{
ydays(x, ...)
years(x, ...)
}
\arguments{
\item{ x}{POSIXct object}
\item{ ...}{Hypothetical arguments of methods for other classes}
}
\value{Numeric
}
\keyword{misc}
